SPDHG algorithm from CCPi optimisation working
